# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CountryDTO {
    private String Code;
    private String name;
}