# src/main/java/ar/edu/utn/frc/tup/lciii/service/SaveCountriesRequest.java
package ar.edu.utn.frc.tup.lciii.service;

public class SaveCountriesRequest {
    private int amountOfCountryToSave;


    public int getAmountOfCountryToSave() {
        return amountOfCountryToSave;
    }

    public void setAmountOfCountryToSave(int amountOfCountryToSave) {
        this.amountOfCountryToSave = amountOfCountryToSave;
    }
}